+++ src/main/java/data/tasks.txt
@@ -1,4 +1,6 @@
-T | 0 | Read Book Monday
-T | 0 | Read paper
+T | 1 | Read Book Monday
+T | 1 | Read paper
 E | 1 | coom | 2019-11-11 | 2020-11-11
 D | 1 | fly | 2012-11-11
+E | 0 | coom | 2019-11-11 | 2020-11-11
+D | 0 | read book | 2011-11-11
