+++ src/main/java/taskmanager/command/Command.java
@@ -1,5 +1,9 @@
 // Command.java (Abstract class)
-package taskmanager;
+package taskmanager.command;
+
+import taskmanager.task.TaskList;
+import taskmanager.ui.Ui;
+import taskmanager.utils.ByteBiteException;
 
 public abstract class Command {
     protected String details;
