+++ src/main/java/taskmanager/command/DeleteCommand.java
@@ -1,5 +1,15 @@
 // DeleteCommand.java
-package taskmanager;
+package taskmanager.command;
+
+import taskmanager.task.TaskList;
+import taskmanager.task.Task;
+import taskmanager.ui.Ui;
+import taskmanager.parser.DateParser;
+import taskmanager.utils.ByteBiteException;
+import taskmanager.utils.InvalidFormatException;
+import taskmanager.utils.TaskNotFoundException;
+import java.time.LocalDate;
+import java.util.ArrayList;
 
 public class DeleteCommand extends Command {
     public DeleteCommand(String details) {
