+++ src/main/java/taskmanager/task/Deadline.java
@@ -1,7 +1,8 @@
 
 // Updated Deadline.java
-package taskmanager;
+package taskmanager.task;
 
+import taskmanager.parser.DateParser;
 import java.time.LocalDate;
 
 public class Deadline extends Task {
