+++ src/main/java/taskmanager/task/Task.java
@@ -1,5 +1,5 @@
 // Task.java
-package taskmanager;
+package taskmanager.task;
 
 public class Task {
    protected String description;
@@ -10,6 +10,7 @@ public class Task {
        this.isDone = false;
    }
 
+   public String getDescription() { return description; }
    public void markAsDone() { isDone = true; }
    public void unmark() { isDone = false; }
    public boolean isDone() { return isDone; }
