+++ src/main/java/taskmanager/utils/ByteBiteException.java
@@ -1,0 +1,8 @@
+// ByteBiteException.java
+package taskmanager.utils;
+
+public class ByteBiteException extends Exception {
+    public ByteBiteException(String message) {
+        super(message);
+    }
+}
