+++ src/main/java/taskmanager/utils/EmptyDescriptionException.java
@@ -1,0 +1,8 @@
+// EmptyDescriptionException.java
+package taskmanager.utils;
+
+public class EmptyDescriptionException extends ByteBiteException {
+    public EmptyDescriptionException(String taskType) {
+        super("The description of a " + taskType + " cannot be empty!");
+    }
+}
