+++ src/main/java/taskmanager/utils/InvalidCommandException.java
@@ -1,0 +1,8 @@
+// InvalidCommandException.java
+package taskmanager.utils;
+
+public class InvalidCommandException extends ByteBiteException {
+    public InvalidCommandException(String command) {
+        super("I don't understand '" + command + "'. Type 'help' for available commands.");
+    }
+}
