+++ src/main/java/taskmanager/utils/InvalidFormatException.java
@@ -1,0 +1,8 @@
+// InvalidFormatException.java
+package taskmanager.utils;
+
+public class InvalidFormatException extends ByteBiteException {
+    public InvalidFormatException(String message) {
+        super(message);
+    }
+}
